Move the admin eligibility answers to their own presenter
The helper method `#display_school` needs to move in to a shared module
as it will be used by the presenters for both policies.

I decided against having the presenters in the `Admin::claim.policy`
namespace as it lead to some weird autoloading/include errors.
